Testing Encoder

+++ NewBasicTemplate2016/src/Subsystems/PIDLogging.h
@@ -29,12 +29,12 @@ protected:
 	double circumfrence;
 
 public:
-	static const int POSITION = 			0b001;
-	static const int POSITION_OFFSET = 		0;
-	static const int VELOCITY = 			0b010;
-	static const int VELOCITY_OFFSET = 		1;
-	static const int DISTANCE = 			0b100;
-	static const int DISTANCE_OFFSET = 		2;
+	static const short int POSITION = 			0b001;
+	static const int POSITION_OFFSET = 			0;
+	static const short int VELOCITY = 			0b010;
+	static const int VELOCITY_OFFSET = 			1;
+	static const short int DISTANCE = 			0b100;
+	static const int DISTANCE_OFFSET = 			2;
 
 	double p, i, d;
